+++ DiffusionNexus.Service/Services/LocalFileMetadataProvider.cs
@@ -1,5 +1,6 @@
 using DiffusionNexus.Service.Classes;
 using DiffusionNexus.Service.Helper;
+using ModelMover.Core.Metadata;
 using System.Security.Cryptography;
 using System.Text.Json;
 
@@ -19,7 +20,7 @@ public class LocalFileMetadataProvider : IModelMetadataProvider
         var fileInfo = new FileInfo(filePath);
         model.SafeTensorFileName = Path.GetFileNameWithoutExtension(filePath);
 
-        var baseName = ExtractBaseName(fileInfo.Name);
+        var baseName = ModelMetadataUtils.ExtractBaseName(fileInfo.Name);
         var directory = fileInfo.Directory;
         var civitaiInfoFile = directory?.GetFiles($"{baseName}.civitai.info").FirstOrDefault();
         var jsonFile = directory?.GetFiles($"{baseName}.json").FirstOrDefault();
@@ -51,7 +52,7 @@ public class LocalFileMetadataProvider : IModelMetadataProvider
             meta.ModelId = modelId.ValueKind switch
             {
                 JsonValueKind.String => modelId.GetString(),
-                JsonValueKind.Number => modelId.GetInt64().ToString(),   // or GetInt32/GetUInt64…
+                JsonValueKind.Number => modelId.GetInt64().ToString(),   // or GetInt32/GetUInt64
                 _ => null
             };
         }
@@ -64,16 +65,16 @@ public class LocalFileMetadataProvider : IModelMetadataProvider
             if (model.TryGetProperty("name", out var name))
                 meta.ModelVersionName = name.GetString();
             if (model.TryGetProperty("type", out var type))
-                meta.ModelType = ParseModelType(type.GetString());
+                meta.ModelType = ModelMetadataUtils.ParseModelType(type.GetString());
             if (model.TryGetProperty("tags", out var tags))
-                meta.Tags = ParseTags(tags);
+                meta.Tags = ModelMetadataUtils.ParseTags(tags);
         }
 
         if (meta.ModelType == DiffusionTypes.UNASSIGNED && root.TryGetProperty("type", out var rootType))
-            meta.ModelType = ParseModelType(rootType.GetString());
+            meta.ModelType = ModelMetadataUtils.ParseModelType(rootType.GetString());
 
         if (meta.Tags.Count == 0 && root.TryGetProperty("tags", out var rootTags))
-            meta.Tags = ParseTags(rootTags);
+            meta.Tags = ModelMetadataUtils.ParseTags(rootTags);
 
         meta.NoMetaData = !meta.HasAnyMetadata;
     }
@@ -87,43 +88,12 @@ public class LocalFileMetadataProvider : IModelMetadataProvider
         if (root.TryGetProperty("sd version", out var ver))
             meta.DiffusionBaseModel = ver.GetString();
         if (root.TryGetProperty("type", out var type))
-            meta.ModelType = ParseModelType(type.GetString());
+            meta.ModelType = ModelMetadataUtils.ParseModelType(type.GetString());
         if (root.TryGetProperty("tags", out var tags))
-            meta.Tags = ParseTags(tags);
+            meta.Tags = ModelMetadataUtils.ParseTags(tags);
 
         meta.NoMetaData = !meta.HasAnyMetadata;
     }
 
-    private static List<string> ParseTags(JsonElement tags)
-    {
-        var result = new List<string>();
-        foreach (var t in tags.EnumerateArray())
-        {
-            if (t.ValueKind == JsonValueKind.String)
-            {
-                var s = t.GetString();
-                if (!string.IsNullOrWhiteSpace(s))
-                    result.Add(s);
-            }
-        }
-        return result;
-    }
 
-    private static DiffusionTypes ParseModelType(string? type)
-    {
-        if (string.IsNullOrWhiteSpace(type))
-            return DiffusionTypes.UNASSIGNED;
-        if (Enum.TryParse(type.Replace(" ", string.Empty), true, out DiffusionTypes dt))
-            return dt;
-        return DiffusionTypes.UNASSIGNED;
-    }
-
-    private static string ExtractBaseName(string fileName)
-    {
-        var known = StaticFileTypes.GeneralExtensions
-            .OrderByDescending(e => e.Length)
-            .FirstOrDefault(e => fileName.EndsWith(e, StringComparison.OrdinalIgnoreCase));
-        return known != null ? fileName[..^known.Length] : fileName;
-    }
 }
-